Update ivi, remove kivi
