--- ivi-v0.7.0/package.json
@@ -1,26 +1,0 @@
-{
-    "private": true,
-    "name": "js-framework-benchmark-ivi",
-    "version": "0.5.0",
-    "description": "Benchmark for ivi",
-    "author": "Stefan Krause",
-    "license": "Apache-2",
-    "homepage": "https://github.com/krausest/js-framework-benchmark",
-    "repository": "https://github.com/krausest/js-framework-benchmark",
-    "keywords": [],
-    "scripts": {
-        "build-dev": "gulp build",
-        "build-prod": "gulp build"
-    },
-    "dependencies": {
-        "ivi": "0.7.0"
-    },
-    "devDependencies": {
-        "del": "2.2.2",
-        "google-closure-compiler-js": "20170218.0.0",
-        "gulp": "3.9.1",
-        "rollup": "0.41.5",
-        "rollup-plugin-node-resolve": "2.0.0",
-        "rollup-plugin-replace": "1.1.1"
-    }
-}
