+++ ivi-v0.8.0-keyed/index.html
@@ -1,12 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
+
 <head>
   <meta charset="UTF-8">
-  <title>ivi v0.7.0</title>
-  <link href="/css/currentStyle.css" rel="stylesheet"/>
+  <title>ivi v0.8.0</title>
+  <link href="/css/currentStyle.css" rel="stylesheet" />
 </head>
+
 <body>
   <div id="main"></div>
   <script src="dist/main.js"></script>
 </body>
+
 </html>
