--- kivi-v1.0.0-rc2/.gitignore
@@ -1,10 +1,0 @@
-build
-dist
-
-node_modules
-
-.DS_Store
-.idea
-.vscode
-.npm-debug
-npm-debug.log
