--- kivi-v1.0.0-rc2/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>kivi v1.0.0-rc0</title>
-  <link href="/css/currentStyle.css" rel="stylesheet"/>
-</head>
-<body>
-  <div id="main"></div>
-  <script src="dist/main.js"></script>
-</body>
-</html>
