--- kivi-v1.0.0-rc2/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es6",
-    "lib": ["es6", "dom"],
-    "outDir": "build/es6",
-    "noImplicitAny": true,
-    "noImplicitReturns": true,
-    "noImplicitThis": true,
-    "noUnusedLocals": true,
-    "strictNullChecks": true,
-    "removeComments": false,
-    "moduleResolution": "node"
-  },
-  "files": [
-    "main.ts"
-  ],
-  "compileOnSave": false
-}
